INTYGFV-11934-2: Modified classes and swapped the hsa usage to hsatk in other infra modules
