--- integration/hsatk-integration/src/main/java/se/inera/intyg/infra/integration/hsatk/model/HealthCareUnitMembersType.java
@@ -1,2 +1,0 @@
-package se.inera.intyg.infra.integration.hsatk.model;public class HealthCareUnitMembersType {
-}
