+++ integration/hsatk-integration/src/main/java/se/inera/intyg/infra/integration/hsatk/model/HospCredentialsForPerson.java
@@ -1,19 +1,38 @@
+/*
+ * Copyright (C) 2020 Inera AB (http://www.inera.se)
+ *
+ * This file is part of sklintyg (https://github.com/sklintyg).
+ *
+ * sklintyg is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * sklintyg is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 package se.inera.intyg.infra.integration.hsatk.model;
 
 import lombok.Data;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Data
 public class HospCredentialsForPerson {
     private String personalIdentityNumber;
-    private List<HealthCareProfessionalLicence> healthCareProfessionalLicence;
+    private List<HealthCareProfessionalLicence> healthCareProfessionalLicence = new ArrayList<>();
     private String personalPrescriptionCode;
-    private List<HCPSpecialityCodes> healthCareProfessionalLicenceSpeciality;
-    private List<NursePrescriptionRight> nursePrescriptionRight;
+    private List<HCPSpecialityCodes> healthCareProfessionalLicenceSpeciality = new ArrayList<>();
+    private List<NursePrescriptionRight> nursePrescriptionRight = new ArrayList<>();
     private String healthcareProfessionalLicenseIdentityNumber;
-    private List<String> educationCode;
-    private List<Restriction> restrictions;
+    private List<String> educationCode = new ArrayList<>();
+    private List<Restriction> restrictions = new ArrayList<>();
     private Boolean feignedPerson;
 
     @Data
