+++ integration/hsatk-integration/src/main/java/se/inera/intyg/infra/integration/hsatk/model/legacy/AbstractVardenhet.java
@@ -93,4 +93,14 @@ public abstract class AbstractVardenhet implements SelectableVardenhet, Comparab
     public int hashCode() {
         return id.hashCode();
     }
+
+    @Override
+    public String getId() {
+        return id;
+    }
+
+    @Override
+    public String getNamn() {
+        return namn;
+    }
 }
