+++ integration/hsatk-integration/src/main/java/se/inera/intyg/infra/integration/hsatk/services/HsatkEmployeeService.java
@@ -1,3 +1,21 @@
+/*
+ * Copyright (C) 2020 Inera AB (http://www.inera.se)
+ *
+ * This file is part of sklintyg (https://github.com/sklintyg).
+ *
+ * sklintyg is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * sklintyg is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 package se.inera.intyg.infra.integration.hsatk.services;
 
 import se.inera.intyg.infra.integration.hsatk.model.PersonInformation;
@@ -5,6 +23,9 @@ import se.inera.intyg.infra.integration.hsatk.model.PersonInformation;
 import java.util.List;
 
 public interface HsatkEmployeeService {
+
+    List<PersonInformation> getEmployee(
+            String personalIdentityNumber, String personHsaId);
 
     List<PersonInformation> getEmployee(
             String personalIdentityNumber, String personHsaId, String profile);
