+++ integration/hsatk-integration/src/main/java/se/inera/intyg/infra/integration/hsatk/services/HsatkEmployeeServiceImpl.java
@@ -1,3 +1,21 @@
+/*
+ * Copyright (C) 2020 Inera AB (http://www.inera.se)
+ *
+ * This file is part of sklintyg (https://github.com/sklintyg).
+ *
+ * sklintyg is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * sklintyg is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 package se.inera.intyg.infra.integration.hsatk.services;
 
 import org.apache.commons.lang3.StringUtils;
@@ -6,14 +24,14 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import se.inera.intyg.infra.integration.hsatk.client.EmployeeClient;
-import se.inera.intyg.infra.integration.hsatk.model.HCPSpecialityCodes;
+import se.inera.intyg.infra.integration.hsatk.exception.HsaServiceCallException;
 import se.inera.intyg.infra.integration.hsatk.model.PersonInformation;
-import se.riv.infrastructure.directory.employee.v2.HealthCareProfessionalLicenceSpecialityType;
-import se.riv.infrastructure.directory.employee.v2.PaTitleType;
+import se.inera.intyg.infra.integration.hsatk.util.HsaTypeConverter;
 import se.riv.infrastructure.directory.employee.v2.PersonInformationType;
 import se.riv.infrastructure.directory.employee.v2.ProfileEnum;
 
 import javax.xml.ws.WebServiceException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -22,8 +40,15 @@ public class HsatkEmployeeServiceImpl implements HsatkEmployeeService {
 
     private static final Logger LOG = LoggerFactory.getLogger(HsatkEmployeeServiceImpl.class);
 
+    HsaTypeConverter hsaTypeConverter = new HsaTypeConverter();
+
     @Autowired
     private EmployeeClient employeeClient;
+
+    @Override
+    public List<PersonInformation> getEmployee(String personalIdentityNumber, String personHsaId) {
+        return getEmployee(personalIdentityNumber, personHsaId, null);
+    }
 
     @Override
     public List<PersonInformation> getEmployee(String personalIdentityNumber, String personHsaId, String profile) {
@@ -32,55 +57,18 @@ public class HsatkEmployeeServiceImpl implements HsatkEmployeeService {
         if (StringUtils.isNotEmpty(profile)) {
             profileEnum = ProfileEnum.fromValue(profile);
         }
+
         try {
             personInformationTypeList = employeeClient.getEmployee(personalIdentityNumber, personHsaId, profileEnum);
-        } catch (Exception e) {
+        } catch (IllegalArgumentException e) {
             LOG.error(e.getMessage());
             throw new WebServiceException(e.getMessage());
+        } catch (HsaServiceCallException exception) {
+            LOG.warn("HsaServiceCallException thrown: {}", exception);
+            personInformationTypeList = new ArrayList<>();
         }
 
-        return personInformationTypeList.stream().map(this::toPersonInformation).collect(Collectors.toList());
+        return personInformationTypeList.stream().map(hsaTypeConverter::toPersonInformation).collect(Collectors.toList());
     }
 
-    public PersonInformation toPersonInformation(PersonInformationType personInformationType) {
-        PersonInformation personInformation = new PersonInformation();
-        personInformation.setAge(personInformationType.getAge());
-        personInformation.setFeignedPerson(personInformationType.isFeignedPerson());
-        personInformation.setGender(personInformationType.getGender());
-        personInformation.setGivenName(personInformationType.getGivenName());
-        personInformation.setHealthCareProfessionalLicence(personInformationType.getHealthCareProfessionalLicence());
-        personInformation.setHealthCareProfessionalLicenceSpeciality(personInformationType.getHealthCareProfessionalLicenceSpeciality()
-                .stream().map(this::toHCPSpecialityCodes).collect(Collectors.toList()));
-        personInformation.setMiddleAndSurName(personInformationType.getMiddleAndSurName());
-        personInformation.setPaTitle(personInformationType.getPaTitle().stream().map(this::toPaTitleType).collect(Collectors.toList()));
-        personInformation.setPersonEndDate(personInformationType.getPersonEndDate());
-        personInformation.setPersonHsaId(personInformationType.getPersonHsaId());
-        personInformation.setPersonStartDate(personInformationType.getPersonStartDate());
-        personInformation.setProtectedPerson(personInformationType.isProtectedPerson());
-        personInformation.setSpecialityCode(personInformationType.getSpecialityCode());
-        personInformation.setSpecialityName(personInformationType.getSpecialityName());
-        return personInformation;
-    }
-
-    private HCPSpecialityCodes toHCPSpecialityCodes(
-            HealthCareProfessionalLicenceSpecialityType healthCareProfessionalLicenceSpecialityType) {
-
-        HCPSpecialityCodes hcpSpecialityCodes = new HCPSpecialityCodes();
-
-        hcpSpecialityCodes.setHealthCareProfessionalLicenceCode(
-                healthCareProfessionalLicenceSpecialityType.getHealthCareProfessionalLicence());
-        hcpSpecialityCodes.setSpecialityCode(healthCareProfessionalLicenceSpecialityType.getSpecialityCode());
-        hcpSpecialityCodes.setSpecialityName(healthCareProfessionalLicenceSpecialityType.getSpecialityName());
-
-        return hcpSpecialityCodes;
-    }
-
-    private PersonInformation.PaTitle toPaTitleType(PaTitleType paTitleType) {
-
-        PersonInformation.PaTitle paTitle = new PersonInformation.PaTitle();
-
-        paTitle.setPaTitleCode(paTitleType.getPaTitleCode());
-        paTitle.setPaTitleName(paTitleType.getPaTitleName());
-        return paTitle;
-    }
 }
