+++ integration/hsatk-integration/src/main/java/se/inera/intyg/infra/integration/hsatk/services/HsatkOrganizationServiceImpl.java
@@ -1,3 +1,21 @@
+/*
+ * Copyright (C) 2020 Inera AB (http://www.inera.se)
+ *
+ * This file is part of sklintyg (https://github.com/sklintyg).
+ *
+ * sklintyg is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * sklintyg is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 package se.inera.intyg.infra.integration.hsatk.services;
 
 import org.apache.commons.lang3.StringUtils;
@@ -7,12 +25,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import se.inera.intyg.infra.integration.hsatk.client.OrganizationClient;
 import se.inera.intyg.infra.integration.hsatk.exception.HsaServiceCallException;
-import se.inera.intyg.infra.integration.hsatk.model.*;
-import se.riv.infrastructure.directory.organization.gethealthcareunitmembersresponder.v2.HealthCareProviderType;
-import se.riv.infrastructure.directory.organization.gethealthcareunitmembersresponder.v2.HealthCareUnitMemberType;
-import se.riv.infrastructure.directory.organization.gethealthcareunitmembersresponder.v2.HealthCareUnitMembersType;
-import se.riv.infrastructure.directory.organization.gethealthcareunitresponder.v2.HealthCareUnitType;
-import se.riv.infrastructure.directory.organization.getunitresponder.v2.*;
+import se.inera.intyg.infra.integration.hsatk.model.HealthCareProvider;
+import se.inera.intyg.infra.integration.hsatk.model.HealthCareUnit;
+import se.inera.intyg.infra.integration.hsatk.model.HealthCareUnitMembers;
+import se.inera.intyg.infra.integration.hsatk.model.Unit;
+import se.inera.intyg.infra.integration.hsatk.util.HsaTypeConverter;
+import se.riv.infrastructure.directory.organization.getunitresponder.v2.GetUnitType;
+import se.riv.infrastructure.directory.organization.getunitresponder.v2.ProfileEnum;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -23,6 +42,8 @@ public class HsatkOrganizationServiceImpl implements HsatkOrganizationService {
 
     private static final Logger LOG = LoggerFactory.getLogger(HsatkAuthorizationManagementServiceImpl.class);
 
+    HsaTypeConverter hsaTypeConverter = new HsaTypeConverter();
+
     @Autowired
     OrganizationClient organizationClient;
 
@@ -32,11 +53,11 @@ public class HsatkOrganizationServiceImpl implements HsatkOrganizationService {
 
         try {
             healthCareProviderList = organizationClient.getHealthCareProvider(healthCareProviderHsaId, healthCareProviderOrgNo)
-                    .stream().map(this::toHealthCareProvider).collect(Collectors.toList());
+                    .stream().map(hsaTypeConverter::toHealthCareProvider).collect(Collectors.toList());
         } catch (HsaServiceCallException e) {
             LOG.error("");
         } catch (Exception e) {
-            LOG.error("");
+            LOG.error("Unexpected error occured: {}");
         }
         return healthCareProviderList;
     }
@@ -46,7 +67,7 @@ public class HsatkOrganizationServiceImpl implements HsatkOrganizationService {
         HealthCareUnit healthCareUnit = new HealthCareUnit();
 
         try {
-            healthCareUnit = toHealthCareUnit(organizationClient.getHealthCareUnit(healthCareUnitMemberHsaId));
+            healthCareUnit = hsaTypeConverter.toHealthCareUnit(organizationClient.getHealthCareUnit(healthCareUnitMemberHsaId));
         } catch (HsaServiceCallException e) {
             LOG.error("");
         } catch (Exception e) {
@@ -59,7 +80,8 @@ public class HsatkOrganizationServiceImpl implements HsatkOrganizationService {
     public HealthCareUnitMembers getHealthCareUnitMembers(String healtCareUnitHsaId) {
         HealthCareUnitMembers healthCareUnitMembers = new HealthCareUnitMembers();
         try {
-            healthCareUnitMembers = toHealthCareUnitMembers(organizationClient.getHealthCareUnitMembers(healtCareUnitHsaId));
+            healthCareUnitMembers = hsaTypeConverter.toHealthCareUnitMembers(
+                    organizationClient.getHealthCareUnitMembers(healtCareUnitHsaId));
         } catch (HsaServiceCallException e) {
             LOG.error("");
         } catch (Exception e) {
@@ -80,7 +102,7 @@ public class HsatkOrganizationServiceImpl implements HsatkOrganizationService {
             profileEnum = ProfileEnum.fromValue(profile);
         }
         try {
-             unit = toUnit(organizationClient.getUnit(unitHsaId, profileEnum));
+            unit = hsaTypeConverter.toUnit(organizationClient.getUnit(unitHsaId, profileEnum));
 
         } catch (HsaServiceCallException e) {
             LOG.error("Failed to get Unit from HSA: {}", e.getLocalizedMessage());
@@ -91,169 +113,4 @@ public class HsatkOrganizationServiceImpl implements HsatkOrganizationService {
         return unit;
     }
 
-    private HealthCareProvider toHealthCareProvider(se.riv.infrastructure.directory.organization.gethealthcareproviderresponder.v1.HealthCareProviderType healthCareProviderType) {
-        HealthCareProvider healthCareProvider = new HealthCareProvider();
-
-        healthCareProvider.setArchivedHealthCareProvider(healthCareProviderType.isArchivedHealthCareProvider());
-        healthCareProvider.setFeignedHealthCareProvider(healthCareProviderType.isFeignedHealthCareProvider());
-        healthCareProvider.setHealthCareProviderEndDate(healthCareProviderType.getHealthCareProviderEndDate());
-        healthCareProvider.setHealthCareProviderHsaId(healthCareProviderType.getHealthCareProviderHsaId());
-        healthCareProvider.setHealthCareProviderName(healthCareProviderType.getHealthCareProviderName());
-        healthCareProvider.setHealthCareProviderOrgNo(healthCareProviderType.getHealthCareProviderOrgNo());
-        healthCareProvider.setHealthCareProviderStartDate(healthCareProviderType.getHealthCareProviderStartDate());
-
-
-        return healthCareProvider;
-    }
-
-    private HealthCareUnit toHealthCareUnit(HealthCareUnitType healthCareUnitType) {
-        HealthCareUnit healthCareUnit = new HealthCareUnit();
-
-        healthCareUnit.setArchivedHealthCareProvider(healthCareUnitType.isArchivedHealthCareProvider());
-        healthCareUnit.setArchivedHealthCareUnit(healthCareUnitType.isArchivedHealthCareUnit());
-        healthCareUnit.setArchivedHealthCareUnitMember(healthCareUnitType.isArchivedHealthCareUnitMember());
-        healthCareUnit.setFeignedHealthCareProvider(healthCareUnitType.isFeignedHealthCareProvider());
-        healthCareUnit.setFeignedHealthCareUnit(healthCareUnitType.isFeignedHealthCareUnit());
-        healthCareUnit.setFeignedHealthCareUnitMember(healthCareUnitType.isFeignedHealthCareUnitMember());
-        healthCareUnit.setHealthCareProviderEndDate(healthCareUnitType.getHealthCareProviderEndDate());
-        healthCareUnit.setHealthCareProviderHsaId(healthCareUnitType.getHealthCareProviderHsaId());
-        healthCareUnit.setHealthCareProviderName(healthCareUnitType.getHealthCareProviderName());
-        healthCareUnit.setHealthCareProviderOrgNo(healthCareUnitType.getHealthCareProviderOrgNo());
-        healthCareUnit.setHealthCareProviderPublicName(healthCareUnitType.getHealthCareProviderPublicName());
-        healthCareUnit.setHealthCareProviderStartDate(healthCareUnitType.getHealthCareProviderStartDate());
-        healthCareUnit.setHealthCareUnitEndDate(healthCareUnitType.getHealthCareUnitEndDate());
-        healthCareUnit.setHealthCareUnitHsaId(healthCareUnitType.getHealthCareUnitHsaId());
-        healthCareUnit.setHealthCareUnitMemberEndDate(healthCareUnitType.getHealthCareUnitMemberEndDate());
-        healthCareUnit.setHealthCareUnitMemberHsaId(healthCareUnitType.getHealthCareUnitMemberHsaId());
-        healthCareUnit.setHealthCareUnitMemberName(healthCareUnitType.getHealthCareUnitMemberName());
-        healthCareUnit.setHealthCareUnitMemberPublicName(healthCareUnitType.getHealthCareUnitMemberPublicName());
-        healthCareUnit.setHealthCareUnitMemberStartDate(healthCareUnitType.getHealthCareUnitMemberStartDate());
-        healthCareUnit.setHealthCareUnitName(healthCareUnitType.getHealthCareUnitName());
-        healthCareUnit.setHealthCareUnitPublicName(healthCareUnitType.getHealthCareUnitPublicName());
-        healthCareUnit.setHealthCareUnitStartDate(healthCareUnitType.getHealthCareUnitStartDate());
-        healthCareUnit.setUnitIsHealthCareUnit(healthCareUnitType.isUnitIsHealthCareUnit());
-
-        return healthCareUnit;
-    }
-
-    private HealthCareUnitMembers toHealthCareUnitMembers(HealthCareUnitMembersType healthCareUnitMembersType) {
-        HealthCareUnitMembers healthCareUnitMembers = new HealthCareUnitMembers();
-
-        healthCareUnitMembers.setArchivedHealthCareUnit(healthCareUnitMembersType.isArchivedHealthCareUnit());
-        healthCareUnitMembers.setFeignedHealthCareUnit(healthCareUnitMembersType.isFeignedHealthCareUnit());
-        if (healthCareUnitMembersType.getHealthCareUnitMember() != null) {
-            healthCareUnitMembers.setHealthCareUnitMember(healthCareUnitMembersType.getHealthCareUnitMember()
-                    .stream().map(this::toHealthCareUnitMember).collect(Collectors.toList()));
-        }
-        healthCareUnitMembers.setHealthCareUnitHsaId(healthCareUnitMembersType.getHealthCareUnitHsaId());
-        healthCareUnitMembers.setHealthCareUnitEndDate(healthCareUnitMembersType.getHealthCareUnitEndDate());
-        healthCareUnitMembers.setHealthCareProvider(toHealthCareProvider(healthCareUnitMembersType.getHealthCareProvider()));
-        healthCareUnitMembers.setHealthCareUnitName(healthCareUnitMembersType.getHealthCareUnitName());
-        healthCareUnitMembers.setHealthCareUnitPrescriptionCode(healthCareUnitMembersType.getHealthCareUnitPrescriptionCode());
-        healthCareUnitMembers.setHealthCareUnitPublicName(healthCareUnitMembersType.getHealthCareUnitPublicName());
-        healthCareUnitMembers.setHealthCareUnitStartDate(healthCareUnitMembersType.getHealthCareUnitStartDate());
-        if (healthCareUnitMembersType.getPostalAddress() != null) {
-            healthCareUnitMembers.setPostalAddress(healthCareUnitMembersType.getPostalAddress().getAddressLine());
-        }
-        healthCareUnitMembers.setPostalCode(healthCareUnitMembersType.getPostalCode());
-        healthCareUnitMembers.setTelephoneNumber(healthCareUnitMembersType.getTelephoneNumber());
-
-
-        return healthCareUnitMembers;
-    }
-
-    private HealthCareUnitMember toHealthCareUnitMember(HealthCareUnitMemberType healthCareUnitMemberType) {
-        HealthCareUnitMember healthCareUnitMember = new HealthCareUnitMember();
-
-        healthCareUnitMember.setArchivedHealthCareUnitMember(healthCareUnitMemberType.isArchivedHealthCareUnitMember());
-        healthCareUnitMember.setFeignedHealthCareUnitMember(healthCareUnitMemberType.isFeignedHealthCareUnitMember());
-        healthCareUnitMember.setHealthCareUnitMemberEndDate(healthCareUnitMemberType.getHealthCareUnitMemberEndDate());
-        healthCareUnitMember.setHealthCareUnitMemberHsaId(healthCareUnitMemberType.getHealthCareUnitMemberHsaId());
-        healthCareUnitMember.setHealthCareUnitMemberName(healthCareUnitMemberType.getHealthCareUnitMemberName());
-        if (healthCareUnitMemberType.getHealthCareUnitMemberpostalAddress() != null) {
-            healthCareUnitMember.setHealthCareUnitMemberpostalAddress(healthCareUnitMemberType.getHealthCareUnitMemberpostalAddress().getAddressLine());
-        }
-        healthCareUnitMember.setHealthCareUnitMemberpostalCode(healthCareUnitMemberType.getHealthCareUnitMemberpostalCode());
-        healthCareUnitMember.setHealthCareUnitMemberPrescriptionCode(healthCareUnitMemberType.getHealthCareUnitMemberPrescriptionCode());
-        healthCareUnitMember.setHealthCareUnitMemberPublicName(healthCareUnitMemberType.getHealthCareUnitMemberPublicName());
-        healthCareUnitMember.setHealthCareUnitMemberStartDate(healthCareUnitMemberType.getHealthCareUnitMemberStartDate());
-        healthCareUnitMember.setHealthCareUnitMemberTelephoneNumber(healthCareUnitMemberType.getHealthCareUnitMemberTelephoneNumber());
-
-        return healthCareUnitMember;
-    }
-
-    private HealthCareProvider toHealthCareProvider(HealthCareProviderType healthCareProviderType) {
-        HealthCareProvider healthCareProvider = new HealthCareProvider();
-
-        healthCareProvider.setArchivedHealthCareProvider(healthCareProviderType.isArchivedHealthCareProvider());
-        healthCareProvider.setFeignedHealthCareProvider(healthCareProviderType.isFeignedHealthCareProvider());
-        healthCareProvider.setHealthCareProviderEndDate(healthCareProviderType.getHealthCareProviderEndDate());
-        healthCareProvider.setHealthCareProviderHsaId(healthCareProviderType.getHealthCareProviderHsaId());
-        healthCareProvider.setHealthCareProviderName(healthCareProviderType.getHealthCareProviderName());
-        healthCareProvider.setHealthCareProviderOrgNo(healthCareProviderType.getHealthCareProviderOrgNo());
-        healthCareProvider.setHealthCareProviderStartDate(healthCareProviderType.getHealthCareProviderStartDate());
-
-        return healthCareProvider;
-    }
-
-    private Unit toUnit(UnitType unitType) {
-        Unit unit = new Unit();
-
-        if (unitType.getBusinessClassification() != null) {
-            unit.setBusinessClassification(unitType.getBusinessClassification()
-                    .stream().map(this::toBusinessClassification).collect(Collectors.toList()));
-        }
-        unit.setBusinessType(unitType.getBusinessType());
-        unit.setCareType(unitType.getCareType());
-        unit.setCountyCode(unitType.getCountyCode());
-        unit.setCountyName(unitType.getCountyName());
-        unit.setFeignedUnit(unitType.isFeignedUnit());
-        if (unitType.getGeographicalCoordinatesRt90() != null) {
-            unit.setGeographicalCoordinatesRt90(toRt90(unitType.getGeographicalCoordinatesRt90()));
-        }
-        if (unitType.getGeographicalCoordinatesSWEREF99() != null) {
-            unit.setGeographicalCoordinatesSWEREF99(toSWEREF99(unitType.getGeographicalCoordinatesSWEREF99()));
-        }
-        unit.setLocation(unitType.getLocation());
-        unit.setManagement(unitType.getManagement());
-        unit.setMunicipalityCode(unitType.getMunicipalityCode());
-        unit.setMunicipalityName(unitType.getMunicipalityName());
-        if (unitType.getPostalAddress() != null) {
-            unit.setPostalAddress(unitType.getPostalAddress().getAddressLine());
-        }
-        unit.setPostalCode(unitType.getPostalCode());
-        unit.setUnitEndDate(unitType.getUnitEndDate());
-        unit.setUnitHsaId(unitType.getUnitHsaId());
-        unit.setUnitName(unitType.getUnitName());
-        unit.setUnitEndDate(unitType.getUnitEndDate());
-
-        return unit;
-    }
-
-    private Unit.GeoCoordRt90 toRt90(GeoCoordRt90Type geoCoordRt90Type) {
-        Unit.GeoCoordRt90 geoCoordRt90 = new Unit.GeoCoordRt90();
-
-        geoCoordRt90.setXCoordinate(geoCoordRt90Type.getXCoordinate());
-        geoCoordRt90.setYCoordinate(geoCoordRt90Type.getYCoordinate());
-
-        return geoCoordRt90;
-    }
-
-    private Unit.GeoCoordSWEREF99 toSWEREF99(GeoCoordSWEREF99Type geoCoordSWEREF99Type) {
-        Unit.GeoCoordSWEREF99 geoCoordSWEREF99 = new Unit.GeoCoordSWEREF99();
-
-        geoCoordSWEREF99.setECoordinate(geoCoordSWEREF99Type.getECoordinate());
-        geoCoordSWEREF99.setNCoordinate(geoCoordSWEREF99Type.getNCoordinate());
-
-        return geoCoordSWEREF99;
-    }
-
-    private Unit.BusinessClassification toBusinessClassification(BusinessClassificationType businessClassificationType) {
-        Unit.BusinessClassification businessClassification = new Unit.BusinessClassification();
-
-        businessClassification.setBusinessClassificationCode(businessClassificationType.getBusinessClassificationCode());
-        businessClassification.setBusinessClassificationName(businessClassificationType.getBusinessClassificationName());
-
-        return businessClassification;
-    }
 }
