+++ integration/hsatk-integration/src/main/java/se/inera/intyg/infra/integration/hsatk/stub/BootstrapBean.java
@@ -1,3 +1,21 @@
+/*
+ * Copyright (C) 2020 Inera AB (http://www.inera.se)
+ *
+ * This file is part of sklintyg (https://github.com/sklintyg).
+ *
+ * sklintyg is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * sklintyg is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 package se.inera.intyg.infra.integration.hsatk.stub;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -33,12 +51,6 @@ public class BootstrapBean {
             addCareProvider(resource);
         }
 
-        files = getResourceListing("bootstrap-ws-careprovider/*.json");
-
-        for (Resource resource : files) {
-            addCareProvider(resource);
-        }
-
         files = getResourceListing("bootstrap-person/*.json");
 
         for (Resource resource : files) {
@@ -46,19 +58,12 @@ public class BootstrapBean {
             addCredentialInformation(resource);
         }
 
-        files = getResourceListing("bootstrap-ws-person/*.json");
+        files = getResourceListing("bootstrap-hospperson/*.json");
 
         for (Resource resource : files) {
             addHsaPerson(resource);
             addCredentialInformation(resource);
         }
-
-        /*files = getResourceListing("bootstrap-hospperson/*.json");
-
-        for (Resource resource : files) {
-            addHsaPerson(resource);
-            addCredentialInformation(resource);
-        }*/
     }
 
     private List<Resource> getResourceListing(String classpathResourcePath) {
