+++ integration/hsatk-integration/src/main/java/se/inera/intyg/infra/integration/hsatk/stub/GetHospCredentialsForPersonResponderStub.java
@@ -1,3 +1,21 @@
+/*
+ * Copyright (C) 2020 Inera AB (http://www.inera.se)
+ *
+ * This file is part of sklintyg (https://github.com/sklintyg).
+ *
+ * sklintyg is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * sklintyg is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 package se.inera.intyg.infra.integration.hsatk.stub;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,13 +26,13 @@ import se.riv.infrastructure.directory.authorizationmanagement.gethospcredential
 import se.riv.infrastructure.directory.authorizationmanagement.v2.IIType;
 import se.riv.infrastructure.directory.authorizationmanagement.v2.RestrictionType;
 
-public class GetHospCredentialsForPersonResponderStub implements GetHospCredentialsForPersonResponderInterface
-{
+public class GetHospCredentialsForPersonResponderStub implements GetHospCredentialsForPersonResponderInterface {
     @Autowired
     HsaServiceStub hsaServiceStub;
 
     @Override
-    public GetHospCredentialsForPersonResponseType getHospCredentialsForPerson(String logicalAddress, GetHospCredentialsForPersonType parameters) {
+    public GetHospCredentialsForPersonResponseType getHospCredentialsForPerson(String logicalAddress,
+                                                                               GetHospCredentialsForPersonType parameters) {
 
         GetHospCredentialsForPersonResponseType response = new GetHospCredentialsForPersonResponseType();
         HsaPerson hsaPerson = hsaServiceStub.getHsaPerson(parameters.getPersonalIdentityNumber());
