+++ integration/hsatk-integration/src/main/java/se/inera/intyg/infra/integration/hsatk/stub/HsaServiceStub.java
@@ -1,9 +1,28 @@
+/*
+ * Copyright (C) 2020 Inera AB (http://www.inera.se)
+ *
+ * This file is part of sklintyg (https://github.com/sklintyg).
+ *
+ * sklintyg is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * sklintyg is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 package se.inera.intyg.infra.integration.hsatk.stub;
 
 import org.springframework.stereotype.Service;
 import se.inera.intyg.infra.integration.hsatk.stub.model.*;
 
 import java.time.LocalDateTime;
+import java.time.ZoneId;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -21,7 +40,7 @@ public class HsaServiceStub {
     private Map<String, CareUnitStub> careUnitMap = new HashMap<>();
     private Map<String, SubUnit> subUnitMap = new HashMap<>();
 
-    private LocalDateTime lastHospUpdate = LocalDateTime.now();
+    private LocalDateTime lastHospUpdate = LocalDateTime.now(ZoneId.systemDefault());
 
     public void addCredentialInformation(CredentialInformation credentialInformation) {
         credentialInformationMap.put(credentialInformation.getHsaId(), credentialInformation);
@@ -145,7 +164,7 @@ public class HsaServiceStub {
     }
 
     public void resetHospLastUpdate() {
-        lastHospUpdate = LocalDateTime.now();
+        lastHospUpdate = LocalDateTime.now(ZoneId.systemDefault());
     }
 
 }
