+++ integration/hsatk-integration/src/main/java/se/inera/intyg/infra/integration/hsatk/stub/model/CareUnitStub.java
@@ -26,6 +26,7 @@ import java.util.List;
 /**
  * @author andreaskaltenbach
  */
+
 @Data
 public class CareUnitStub extends AbstractCareUnitStub {
