+++ integration/hsatk-integration/src/main/java/se/inera/intyg/infra/integration/hsatk/stub/model/GeoCoordConverter.java
@@ -1,3 +1,21 @@
+/*
+ * Copyright (C) 2020 Inera AB (http://www.inera.se)
+ *
+ * This file is part of sklintyg (https://github.com/sklintyg).
+ *
+ * sklintyg is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * sklintyg is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 package se.inera.intyg.infra.integration.hsatk.stub.model;
 
 import se.inera.intyg.infra.integration.hsatk.model.Unit;
@@ -14,13 +32,13 @@ public class GeoCoordConverter {
         return geoCoordRt90;
     }
 
-    public Unit.GeoCoordSWEREF99 toSWEREF99(GeoCoordSWEREF99Type geoCoordSWEREF99Type) {
-        Unit.GeoCoordSWEREF99 geoCoordSWEREF99 = new Unit.GeoCoordSWEREF99();
+    public Unit.GeoCoordSweref99 toSweref99(GeoCoordSWEREF99Type geoCoordSweref99Type) {
+        Unit.GeoCoordSweref99 geoCoordSweref99 = new Unit.GeoCoordSweref99();
 
-        geoCoordSWEREF99.setECoordinate(geoCoordSWEREF99Type.getECoordinate());
-        geoCoordSWEREF99.setNCoordinate(geoCoordSWEREF99Type.getNCoordinate());
+        geoCoordSweref99.setECoordinate(geoCoordSweref99Type.getECoordinate());
+        geoCoordSweref99.setNCoordinate(geoCoordSweref99Type.getNCoordinate());
 
-        return geoCoordSWEREF99;
+        return geoCoordSweref99;
     }
 
 }
